SMART CAR PARKING SYSTEM BASED ON IOT
The main aim of this project is to create a completely automated carparking system
with minimal human interference. The fully automated smart car parking system is
rudimental and does not require heavy lines of code nor expensive equipment. It is a
simple circuit built for the exact need of purpose. This automated system Is used to
find the vacancy in parking spaces available and navigate the driver to reach the desired space
using visuals and in an effective manner, thus reducing search time. This system is required
for malls, multistorey parking structures, IT hubs and parking facilities. Smart Parking
System that uses cluster based algorithm to handle the dynamic parking allocation problem
.In our project we propose a smart and car parking model that will help the user in booking
their parking spaces before hand.